+++ test-data/blocks.py
@@ -1,0 +1,12 @@
+#!/usr/bin/env python3
+
+# Print out the unicde block range
+
+s = ""
+for b in range(0x2580, 0x25A0):
+    s += chr(b) + " "
+
+print(s)
+print(s)
+print(s)
+