+++ wezterm-gui/src/termwindow/render.rs
@@ -1,5 +1,5 @@
 use crate::glium::texture::SrgbTexture2d;
-use crate::glyphcache::{CachedGlyph, GlyphCache};
+use crate::glyphcache::{BlockKey, CachedGlyph, GlyphCache};
 use crate::shapecache::*;
 use crate::termwindow::{BorrowedShapeCacheKey, MappedQuads, RenderState, ScrollHit, ShapedInfo};
 use ::window::bitmaps::atlas::OutOfTextureSpace;
@@ -779,6 +779,25 @@ impl super::TermWindow {
                         continue;
                     }
 
+                    if self.config.custom_block_glyphs {
+                        if let Some(block) = BlockKey::from_cell(&params.line.cells()[cell_idx]) {
+                            self.populate_block_quad(
+                                block,
+                                gl_state,
+                                quads,
+                                cell_idx,
+                                &params,
+                                hsv,
+                                cursor_shape,
+                                glyph_color,
+                                underline_color,
+                                bg_color,
+                                white_space,
+                            )?;
+                            continue;
+                        }
+                    }
+
                     let texture = glyph
                         .texture
                         .as_ref()
@@ -885,6 +904,51 @@ impl super::TermWindow {
             );
             quad.set_cursor_color(params.cursor_border_color);
         }
+
+        Ok(())
+    }
+
+    pub fn populate_block_quad(
+        &self,
+        block: BlockKey,
+        gl_state: &RenderState,
+        quads: &mut MappedQuads,
+        cell_idx: usize,
+        params: &RenderScreenLineOpenGLParams,
+        hsv: Option<config::HsbTransform>,
+        cursor_shape: Option<CursorShape>,
+        glyph_color: Color,
+        underline_color: Color,
+        bg_color: Color,
+        white_space: TextureRect,
+    ) -> anyhow::Result<()> {
+        let sprite = gl_state
+            .glyph_cache
+            .borrow_mut()
+            .cached_block(block)?
+            .texture_coords();
+
+        let mut quad =
+            match quads.cell(cell_idx + params.pos.left, params.line_idx + params.pos.top) {
+                Ok(quad) => quad,
+                Err(_) => return Ok(()),
+            };
+
+        quad.set_hsv(hsv);
+        quad.set_fg_color(glyph_color);
+        quad.set_underline_color(underline_color);
+        quad.set_bg_color(bg_color);
+        quad.set_texture(sprite);
+        quad.set_texture_adjust(0., 0., 0., 0.);
+        quad.set_underline(white_space);
+        quad.set_has_color(false);
+        quad.set_cursor(
+            gl_state
+                .util_sprites
+                .cursor_sprite(cursor_shape)
+                .texture_coords(),
+        );
+        quad.set_cursor_color(params.cursor_border_color);
 
         Ok(())
     }